新增OWNER_EMAIL的env變數，控制可以看見我的模型，以及對相關APICRUD的權限

+++ server/api/myModels/ModelFinishInfo/[id].delete.ts
@@ -1,5 +1,6 @@
 import { prisma } from '@/server/utils/prisma'
 export default defineEventHandler(async (event)=>{
+    await requireOwner(event)
     const { id } = event.context.params as any //是FinishInfo的id
 
     const finishInfo =await prisma.modelFinishInfo.delete({
