+++ server/api/myModels/MyModel/[myModelId].delete.ts
@@ -1,5 +1,6 @@
 import { prisma } from '@/server/utils/prisma'
 export default defineEventHandler(async (event)=>{
+    await requireOwner(event)
     const { myModelId } = event.context.params as any
 
     const myModel =await prisma.myModel.delete({
